sketch-ranker/requirements.txt
pandas
pillow
selenium
webdriver-manager
